chore(docs.yml): add linting step to the workflow chore(Makefile): add linting target for Dockerfile refactor(Dockerfile): optimize build process by skipping build if storybook-static directory exists
The changes in the `.github/workflows/docs.yml` file add a new step to the workflow called "Lint" which runs the command `make lint-docker-storybook`. This step is responsible for linting the Dockerfile used for building Storybook.

The changes in the `Makefile` add a new target called `lint-docker-storybook` which runs the command `docker run --rm -i hadolint/hadolint hadolint --ignore DL3018 - < ${STORYBOOK_DOCKERFILE}`. This target is used to lint the Dockerfile for Storybook.

The changes in the `Dockerfile` optimize the build process by skipping the build if the `storybook-static` directory already exists. This is done by checking if the directory exists before running the build commands. If the directory exists, the build process is skipped. This helps to improve the build time and avoid unnecessary rebuilds.