feat(docs-workflow): add reusable dev workflow for documentation generation
This commit adds a new workflow file, `docs-workflow.yml`, which defines a reusable development workflow for generating documentation. The workflow is triggered by a `workflow_call` event and accepts several inputs such as `node-version`, `make-file`, `make-lint-task`, `make-build-task`, and `make-test-task`. These inputs allow for customization of the workflow based on the specific project requirements.

The workflow consists of the following steps:
1. Checkout the repository
2. Setup Node.js environment based on the specified `node-version`
3. Run linting using the specified `make-file` and `make-lint-task`
4. Build the documentation using the specified `make-file` and `make-build-task`
5. Run tests using the specified `make-file` and `make-test-task`
6. Set the `VERSION` environment variable based on the contents of the `VERSION` file
7. Log in to GHCR (GitHub Container Registry) using the actor's credentials
8. Publish the documentation by running the `docker-storybook` task

The `docs.yml` file is also modified to use the new `docs-workflow.yml` workflow file for packaging Storybook. The `build-storybook` job is commented out as it is no longer needed.

The `VERSION` file is added to the repository, which contains the current version of the project.

A new `docs.mk` file is added, which defines the Makefile commands for building and packaging Storybook. The `docker-storybook` target builds a Docker image for Storybook using the specified Dockerfile and pushes it to the specified Docker repository.

These changes are made to improve the development workflow for generating documentation and simplify the process of packaging Storybook.