feat(layout): add G2Row export to index.ts in layout package refactor(G2ConfigBuilder): import RequestProps from '@komune-io/g2-utils' in RequestProps.ts to improve code organization and maintainability
The G2Row export is added to the index.ts file in the layout package to make it available for external use. In the G2ConfigBuilder's RequestProps.ts file, the RequestProps type is now imported from '@komune-io/g2-utils' to enhance code organization and maintainability by centralizing the type definition.