feat(AutoForm): add optional section with conditional display
Add a new optional section in the AutoForm JSON schema that can be
conditionally displayed based on the value of a checkbox. This
enhances the form's flexibility by allowing sections to be shown or
hidden dynamically.

fix(AutoForm): correct typo in formSectionsDisplay variable

Correct a typo in the formSectionsDisplay variable name to ensure
proper rendering of form sections.

refactor(Conditions): separate display and message condition evaluation

Refactor condition evaluation by separating display and message
conditions into distinct functions. This improves code readability
and maintainability.

chore(http): export useQueryRequestArray from utils

Add export for useQueryRequestArray in the HTTP utilities to make it
available for use in other parts of the application.